Combined Superstructure into Robot
